graphql/compliance/pyproject.toml
[project]
name = "syntek-graphql-compliance"
version = "1.0.0"
description = "GraphQL operations for GDPR compliance and legal document management"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [
    { name = "Syntek Dev Team", email = "dev@syntek.com" }
]
keywords = ["graphql", "gdpr", "compliance", "legal", "privacy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Framework :: Django",
    "Framework :: Django :: 6.0",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    "django>=6.0.2",
    "strawberry-graphql>=0.291.0",
    "syntek-graphql-core>=1.0.0",
    "syntek-compliance>=1.0.0",
    "syntek-audit>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.9.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.9.1",
    "mypy>=1.14.0",
    "django-stubs>=5.1.1",
]

[project.urls]
Homepage = "https://github.com/syntek/syntek-modules"
Repository = "https://github.com/syntek/syntek-modules"
Documentation = "https://docs.syntek.com/graphql-compliance"
"Bug Tracker" = "https://github.com/syntek/syntek-modules/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["syntek_graphql_compliance"]

[tool.ruff]
line-length = 100
target-version = "py314"
extend-exclude = ["migrations", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DJ",     # flake8-django
    "TID",    # flake8-tidy-imports
    "SIM",    # flake8-simplify
    "TC",     # flake8-type-checking
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "DJ001",  # Django model nullable CharField/TextField
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"*/types/*.py" = ["TC003"]  # typing-only import moved to TYPE_CHECKING block

[tool.ruff.lint.isort]
known-first-party = ["syntek_graphql_compliance", "syntek_graphql_core"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["mypy_django_plugin.main", "strawberry.ext.mypy_plugin"]

[[tool.mypy.overrides]]
module = "apps.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=syntek_graphql_compliance",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]

[tool.coverage.run]
source = ["syntek_graphql_compliance"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "@abc.abstractmethod",
]
